requirements.txt
python-magic>=0.4.22
yara-python>=4.0.5
